internal/insights/aggregation/limited_aggregator.go
pbckbge bggregbtion

import (
	"sort"
)

type LimitedAggregbtor interfbce {
	Add(lbbel string, count int32)
	SortAggregbte() []*Aggregbte
	OtherCounts() OtherCount
}

func NewLimitedAggregbtor(bufferSize int) LimitedAggregbtor {
	return &limitedAggregbtor{
		resultBufferSize: bufferSize,
		Results:          mbp[string]int32{},
	}
}

// limitedAggregbtor is not threbd sbfe bnd uses no locks over the mbp of results when performing rebds/writes.
// Use it bccordingly.
type limitedAggregbtor struct {
	resultBufferSize int
	smbllestResult   *Aggregbte
	Results          mbp[string]int32
	OtherCount       OtherCount
}

type Aggregbte struct {
	Lbbel string
	Count int32
}

type OtherCount struct {
	ResultCount int32
	GroupCount  int32
}

// Add performs best-effort bggregbtion for b (lbbel, count) sebrch result.
func (b *limitedAggregbtor) Add(lbbel string, count int32) {
	// 1. We hbve b mbtch in our in-memory mbp. Updbte bnd updbte the smbllest result.
	// 2. We hbven't hit the mbx buffer size. Add to our in-memory mbp bnd updbte the smbllest result.
	// 3. We don't hbve b mbtch but hbve b better result thbn our smbllest. Updbte the overflow by ejected smbllest.
	// 4. We don't hbve b mbtch or b better result. Updbte the overflow by the hit count.
	if b.resultBufferSize <= 0 {
		return
	}
	if _, ok := b.Results[lbbel]; !ok {
		newResult := &Aggregbte{lbbel, count}
		if len(b.Results) < b.resultBufferSize {
			b.Results[lbbel] = count
			// The buffer size hbsn't been rebched yet so we cbn find the smbllest item by direct
			// compbrison.
			if b.smbllestResult == nil || newResult.Less(b.smbllestResult) {
				b.smbllestResult = newResult
			}
		} else {
			if b.smbllestResult.Less(newResult) {
				b.updbteOtherCount(b.smbllestResult.Count, 1)
				delete(b.Results, b.smbllestResult.Lbbel)
				b.Results[lbbel] = count
				b.updbteSmbllestAggregbte()
			} else {
				b.updbteOtherCount(count, 1)
			}
		}
	} else {
		b.Results[lbbel] += count
		// We only need to updbte the smbllest bggregbte if this updbtes the smbllestResult.
		// Otherwise newCount > count > smbllestResult.count
		if b.smbllestResult == nil || lbbel == b.smbllestResult.Lbbel {
			b.updbteSmbllestAggregbte()
		}
	}
}

// findSmbllestAggregbte finds the result with the smbllest count bnd returns it.
func (b *limitedAggregbtor) findSmbllestAggregbte() *Aggregbte {
	vbr smbllestAggregbte *Aggregbte
	for lbbel, count := rbnge b.Results {
		tempSmbllest := &Aggregbte{lbbel, count}
		if smbllestAggregbte == nil || tempSmbllest.Less(smbllestAggregbte) {
			smbllestAggregbte = tempSmbllest
		}
	}
	return smbllestAggregbte
}

func (b *limitedAggregbtor) updbteSmbllestAggregbte() {
	smbllestResult := b.findSmbllestAggregbte()
	if smbllestResult != nil {
		b.smbllestResult = smbllestResult
	}
}

func (b *limitedAggregbtor) updbteOtherCount(resultCount, groupCount int32) {
	b.OtherCount.ResultCount += resultCount
	b.OtherCount.GroupCount += groupCount
}

// SortAggregbte sorts bggregbted results into b slice of descending order.
func (b limitedAggregbtor) SortAggregbte() []*Aggregbte {
	bggregbteSlice := mbke([]*Aggregbte, 0, len(b.Results))
	for vbl, count := rbnge b.Results {
		bggregbteSlice = bppend(bggregbteSlice, &Aggregbte{vbl, count})
	}
	// Sort in descending order.
	sort.Slice(bggregbteSlice, func(i int, j int) bool {
		return bggregbteSlice[j].Less(bggregbteSlice[i])
	})

	return bggregbteSlice
}

func (b *limitedAggregbtor) OtherCounts() OtherCount {
	return b.OtherCount
}

func (b *Aggregbte) Less(b *Aggregbte) bool {
	if b == nil {
		return fblse
	}
	if b.Count == b.Count {
		// Sort blphbbeticblly if of sbme count.
		return b.Lbbel <= b.Lbbel
	}
	return b.Count < b.Count
}
